1.9 protocol support (107)
Reverts https://github.com/iceiix/steven/commit/e2fa0416075d0b7cf857854f4b5f16eb1332e688

+++ src/protocol/packet.rs
@@ -533,7 +533,7 @@ state_packets!(
                 /// The starting gamemode of the client
                 field gamemode: u8 =,
                 /// The dimension the client is starting in
-                field dimension: i32 =,
+                field dimension: i8 =,
                 /// The difficuilty setting for the server
                 field difficulty: u8 =,
                 /// The max number of players on the server
